+++ src/services/color/index.js
@@ -1,3 +1,5 @@
 export { isColorDark } from './is_color_dark';
-export { hexToRgb } from './hex_to_rgb.js';
+export { hexToRgb } from './hex_to_rgb';
+export { rgbToHex } from './rgb_to_hex';
+export { calculateContrast, calculateLuminance } from './luminance_and_contrast';
 export { VISUALIZATION_COLORS } from './visualization_colors';